Refactor as app grows
